tss2-mu: fix precision warnings tpm2b-types.c
Fixes warnings from Visual Studios cl.exe compiler like:
tpm2-tss/src/tss2-mu/tpm2b-types.c(352): warning C4244: 'function': conversion from 'int' to 'uint16_t', possible loss of data
tpm2-tss/src/tss2-mu/tpm2b-types.c(354): warning C4244: 'function': conversion from 'int' to 'uint16_t', possible loss of data
tpm2-tss/src/tss2-mu/tpm2b-types.c(356): warning C4244: 'function': conversion from 'int' to 'uint16_t', possible loss of data
tpm2-tss/src/tss2-mu/tpm2b-types.c(358): warning C4244: 'function': conversion from 'int' to 'uint16_t', possible loss of data
tpm2-tss/src/tss2-mu/tpm2b-types.c(360): warning C4244: 'function': conversion from 'int' to 'uint16_t', possible loss of data
tpm2-tss/src/tss2-mu/tpm2b-types.c(362): warning C4244: 'function': conversion from 'int' to 'uint16_t', possible loss of data

All the casts are safe, just add them.

+++ src/tss2-mu/tpm2b-types.c
@@ -209,7 +209,7 @@ TSS2_RC Tss2_MU_##type##_Marshal(type const *src, uint8_t buffer[], \
 \
     /* Update the size to the real value */ \
     if (buffer) { \
-        UINT16 t = HOST_TO_BE_16(buffer + local_offset - ptr - 2); \
+        UINT16 t = HOST_TO_BE_16((UINT16)(buffer + local_offset - ptr - 2)); \
         memcpy(ptr, &t, sizeof(t)); \
     } \
 \
